requirements.txt
# Box to Google Earth Pipeline Requirements

# Box SDK for authentication and file operations
boxsdk>=3.9.0,<4.0.0

# Excel file processing
openpyxl>=3.1.2
pandas>=2.1.0

# KML/KMZ generation
simplekml>=1.3.6

# AWS SDK for S3 and CloudFront
boto3>=1.28.0

# Configuration management
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.9.0
flake8>=6.1.0

# Optional: HTTP health check endpoint
flask>=3.0.0
flask-cors>=4.0.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.1

# Database
pymysql>=1.1.0
cryptography>=41.0.0

# Utilities
python-dateutil>=2.8.2

# Debugging
debugpy>=1.8.0


